backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt
qrcode
pillow
python-multipart
reportlab
python-dotenv
psycopg2-binary
